validation added for the accoount number in card microservice

+++ account/src/main/java/com/finaxis/account/service/AccountService.java
@@ -8,5 +8,5 @@ public interface AccountService {
     int modifyAccountDetails(Account account);
     int deleteAccountDetails(int accountId);
     Account getAccountDetails(int accountId);
-    CardDetailsDTO fecthCardDetailsOfUser(int accountRef);
+    CardDetailsDTO fecthCardDetailsOfUser(long accountRef);
 }
